Added ability to look multiple turns ahead.
Autoplayer will now create a copy of the grid to plan out multiple moves
in the future. The number of turns to look ahead can be configured in
the Autoplayer.js code.

Future moves are weighted less depending on how far in the future they
are (weight decreases by half every turn).

To facilitate these changes, a lot of the functions for moving tiles was
moved from the GameManager class to the Grid class.