Fixed overweighting of 3 in a row moves
The autoplayer now tracks which tiles have already been counted as
merged in a move. This prevents it from double counting in certain
situations and overvaluing inefficient moves.

Barring further in depth analysis showing that it is making incorrect
choices, I'm considering it's current move weighting algorithm to be
working as designed.